WIP on proptests
