+++ mqtt/policy/src/core/mod.rs
@@ -843,4 +843,80 @@ pub(crate) mod tests {
             policy.starts_with(input)
         }
     }
+
+    use crate::{Decision, Effect, PolicyBuilder, PolicyDefinition, Request, Statement};
+    use proptest::{collection::vec, prelude::*};
+
+    proptest! {
+        #[test]
+        fn policy_builder_does_not_crash(definition in arb_policy_definition()){
+            let statement = &definition.statements()[0];
+
+            let request = Request::new(
+                &statement.identities()[0],
+                &statement.operations()[0],
+                &statement.resources()[0],
+            ).expect("unable to create a request");
+
+            let expected = match statement.effect() {
+                Effect::Allow => Decision::Allowed,
+                Effect::Deny => Decision::Denied
+            };
+
+            let policy = PolicyBuilder::from_definition(definition)
+                .build()
+                .expect("unable to build policy from definition");
+
+            assert_eq!(policy.evaluate(&request).unwrap(), expected);
+        }
+    }
+
+    prop_compose! {
+        pub fn arb_policy_definition()(
+            statements in vec(arb_statement(), 1..3)
+        ) -> PolicyDefinition {
+            PolicyDefinition {
+                statements
+            }
+        }
+    }
+
+    prop_compose! {
+        pub fn arb_statement()(
+            description in arb_description(),
+            effect in arb_effect(),
+            identities in vec(arb_identity(), 1..3),
+            operations in vec(arb_operation(), 1..3),
+            resources in vec(arb_resource(), 1..3),
+        ) -> Statement {
+            Statement{
+                order: 0,
+                description,
+                effect,
+                identities,
+                operations,
+                resources,
+            }
+        }
+    }
+
+    pub fn arb_effect() -> impl Strategy<Value = Effect> {
+        prop_oneof![Just(Effect::Allow), Just(Effect::Deny)]
+    }
+
+    pub fn arb_description() -> impl Strategy<Value = String> {
+        "\\PC+"
+    }
+
+    pub fn arb_identity() -> impl Strategy<Value = String> {
+        "(\\PC+)|(\\{\\{\\PC+\\}\\})"
+    }
+
+    pub fn arb_operation() -> impl Strategy<Value = String> {
+        "\\PC+"
+    }
+
+    pub fn arb_resource() -> impl Strategy<Value = String> {
+        "\\PC+(/(\\PC+|\\{\\{\\PC+\\}\\}))*"
+    }
 }
